minor bug fixes: mainly, do not check if Host is Ready upon every invokation, since it might enfore an additional query to the I2C/deviceRouter peripheral

+++ Src/GPIO/Communication/SerialDeviceBase.cs
@@ -111,6 +111,13 @@ namespace R2Core.GPIO
 
         public void Delete() {
 
+            if (Deleted) {
+
+                Log.i($"Will not delete {Identifier}, since it's deleted.");
+                return;
+
+            }
+
             Host.DeleteDevice(DeviceId, Node.NodeId);
             Node.RemoveDevice(this);
             Deleted = true;
